move remaining `SwitchPort*` types from common to nexus-types
