+++ common/src/api/external/mod.rs
@@ -2589,74 +2589,6 @@ impl SimpleIdentity for LldpNeighbor {
     }
 }
 
-/// A route configuration for a port settings object.
-#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
-pub struct SwitchPortRouteConfig {
-    /// The port settings object this route configuration belongs to.
-    pub port_settings_id: Uuid,
-
-    /// The interface name this route configuration is assigned to.
-    pub interface_name: Name,
-
-    /// The route's destination network.
-    pub dst: oxnet::IpNet,
-
-    /// The route's gateway address.
-    pub gw: IpAddr,
-
-    /// The VLAN identifier for the route. Use this if the gateway is reachable
-    /// over an 802.1Q tagged L2 segment.
-    pub vlan_id: Option<u16>,
-
-    /// Route RIB priority. Higher priority indicates precedence within and across
-    /// protocols.
-    pub rib_priority: Option<u8>,
-}
-
-/// An IP address configuration for a port settings object.
-#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
-pub struct SwitchPortAddressConfig {
-    /// The port settings object this address configuration belongs to.
-    pub port_settings_id: Uuid,
-
-    /// The id of the address lot block this address is drawn from.
-    pub address_lot_block_id: Uuid,
-
-    /// The IP address and prefix.
-    pub address: oxnet::IpNet,
-
-    /// An optional VLAN ID
-    pub vlan_id: Option<u16>,
-
-    /// The interface name this address belongs to.
-    pub interface_name: Name,
-}
-
-/// An IP address configuration for a port settings object.
-#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
-pub struct SwitchPortAddressView {
-    /// The port settings object this address configuration belongs to.
-    pub port_settings_id: Uuid,
-
-    /// The id of the address lot this address is drawn from.
-    pub address_lot_id: Uuid,
-
-    /// The name of the address lot this address is drawn from.
-    pub address_lot_name: Name,
-
-    /// The id of the address lot block this address is drawn from.
-    pub address_lot_block_id: Uuid,
-
-    /// The IP address and prefix.
-    pub address: oxnet::IpNet,
-
-    /// An optional VLAN ID
-    pub vlan_id: Option<u16>,
-
-    /// The interface name this address belongs to.
-    pub interface_name: Name,
-}
-
 /// Configuration of inbound ICMP allowed by API services.
 #[derive(
     Clone,
