+++ nexus/types/versions/src/initial/networking.rs
@@ -8,7 +8,6 @@
 //! types.
 
 use api_identity::ObjectIdentity;
-use omicron_common::api::external;
 use omicron_common::api::external::{
     AddressLotKind, IdentityMetadata, IdentityMetadataCreateParams, Name,
     NameOrId, ObjectIdentity,
@@ -265,6 +264,74 @@ pub enum SwitchPortGeometry {
 
     /// The port contains four SFP28 links each with one lane.
     Sfp28x4,
+}
+
+/// A route configuration for a port settings object.
+#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
+pub struct SwitchPortRouteConfig {
+    /// The port settings object this route configuration belongs to.
+    pub port_settings_id: Uuid,
+
+    /// The interface name this route configuration is assigned to.
+    pub interface_name: Name,
+
+    /// The route's destination network.
+    pub dst: oxnet::IpNet,
+
+    /// The route's gateway address.
+    pub gw: IpAddr,
+
+    /// The VLAN identifier for the route. Use this if the gateway is reachable
+    /// over an 802.1Q tagged L2 segment.
+    pub vlan_id: Option<u16>,
+
+    /// Route RIB priority. Higher priority indicates precedence within and across
+    /// protocols.
+    pub rib_priority: Option<u8>,
+}
+
+/// An IP address configuration for a port settings object.
+#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
+pub struct SwitchPortAddressConfig {
+    /// The port settings object this address configuration belongs to.
+    pub port_settings_id: Uuid,
+
+    /// The id of the address lot block this address is drawn from.
+    pub address_lot_block_id: Uuid,
+
+    /// The IP address and prefix.
+    pub address: oxnet::IpNet,
+
+    /// An optional VLAN ID
+    pub vlan_id: Option<u16>,
+
+    /// The interface name this address belongs to.
+    pub interface_name: Name,
+}
+
+/// An IP address configuration for a port settings object.
+#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
+pub struct SwitchPortAddressView {
+    /// The port settings object this address configuration belongs to.
+    pub port_settings_id: Uuid,
+
+    /// The id of the address lot this address is drawn from.
+    pub address_lot_id: Uuid,
+
+    /// The name of the address lot this address is drawn from.
+    pub address_lot_name: Name,
+
+    /// The id of the address lot block this address is drawn from.
+    pub address_lot_block_id: Uuid,
+
+    /// The IP address and prefix.
+    pub address: oxnet::IpNet,
+
+    /// An optional VLAN ID
+    pub vlan_id: Option<u16>,
+
+    /// The interface name this address belongs to.
+    pub interface_name: Name,
 }
 
 /// Switch link configuration.
@@ -945,12 +1012,6 @@ pub struct BgpConfig {
 
 // SWITCH PORT SETTINGS (old response type with required BgpPeer.addr)
 
-/// Switch port settings (old version with required BgpPeer.addr).
-// TODO: several fields below embed `external::*` types directly from
-// `omicron-common`, which means their serialized shape is not truly frozen.
-// Once `omicron-common-versions` exists, replace these with version-local
-// copies of the types to ensure the initial version's wire format is
-// immutable.
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct SwitchPortSettings {
     #[serde(flatten)]
@@ -972,13 +1033,13 @@ pub struct SwitchPortSettings {
     pub vlan_interfaces: Vec<SwitchVlanInterfaceConfig>,
 
     /// IP route settings.
-    pub routes: Vec<external::SwitchPortRouteConfig>,
+    pub routes: Vec<SwitchPortRouteConfig>,
 
     /// BGP peer settings.
     pub bgp_peers: Vec<BgpPeer>,
 
     /// Layer 3 IP address settings.
-    pub addresses: Vec<external::SwitchPortAddressView>,
+    pub addresses: Vec<SwitchPortAddressView>,
 }
 
 /// A link configuration for a port settings object.
