+++ nexus/types/versions/src/stronger_bgp_unnumbered_types/networking.rs
@@ -20,10 +20,10 @@
 
 use crate::v2025_11_20_00::networking::{
     AddressConfig, LinkConfigCreate, RouteConfig, SwitchInterfaceConfig,
-    SwitchInterfaceConfigCreate, SwitchPortConfig, SwitchPortConfigCreate,
-    SwitchPortLinkConfig, SwitchPortSettingsGroups, SwitchVlanInterfaceConfig,
+    SwitchInterfaceConfigCreate, SwitchPortAddressView, SwitchPortConfig,
+    SwitchPortConfigCreate, SwitchPortLinkConfig, SwitchPortRouteConfig,
+    SwitchPortSettingsGroups, SwitchVlanInterfaceConfig,
 };
-use omicron_common::api::external;
 use omicron_common::api::external::IdentityMetadata;
 use omicron_common::api::external::IdentityMetadataCreateParams;
 use omicron_common::api::external::Name;
@@ -324,11 +324,6 @@ impl TryFrom<crate::v2026_02_13_01::networking::SwitchPortSettingsCreate>
 /// This structure contains all port settings information in one place. It's a
 /// convenience data structure for getting a complete view of a particular
 /// port's settings.
-// TODO: several fields below embed `external::*` types directly from
-// `omicron-common`, which means their serialized shape is not truly frozen.
-// Once `omicron-common-versions` exists, replace these with version-local
-// copies of the types to ensure the initial version's wire format is
-// immutable.
 #[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
 pub struct SwitchPortSettings {
     #[serde(flatten)]
@@ -350,13 +345,13 @@ pub struct SwitchPortSettings {
     pub vlan_interfaces: Vec<SwitchVlanInterfaceConfig>,
 
     /// IP route settings.
-    pub routes: Vec<external::SwitchPortRouteConfig>,
+    pub routes: Vec<SwitchPortRouteConfig>,
 
     /// BGP peer settings.
     pub bgp_peers: Vec<BgpPeer>,
 
     /// Layer 3 IP address settings.
-    pub addresses: Vec<external::SwitchPortAddressView>,
+    pub addresses: Vec<SwitchPortAddressView>,
 }
 
 impl From<SwitchPortSettings>
